+++ pykechain/models/activity2.py
@@ -12,6 +12,7 @@ from pykechain.enums import ActivityType, ActivityStatus, Category, ActivityClas
 from pykechain.exceptions import NotFoundError, IllegalArgumentError, APIError, MultipleFoundError
 from pykechain.models.input_checks import check_datetime, check_text, check_list_of_text, check_enum, check_user, \
     check_type
+from pykechain.models.representations.component import RepresentationsComponent
 from pykechain.models.tags import TagsMixin
 from pykechain.models.tree_traversal import TreeObject
 from pykechain.models.user import User
@@ -56,8 +57,14 @@ class Activity2(TreeObject, TagsMixin):
         self.start_date = parse_datetime(json.get('start_date'))
         self.due_date = parse_datetime(json.get('due_date'))
         self.assignees_ids = json.get('assignees_ids', [])  # type: List[Text]
+        self._options = json.get('activity_options', {})
 
         self._tags = json.get('tags', [])  # type: List[Text]
+        self._representations_container = RepresentationsComponent(
+            self,
+            self._options.get('representations', {}),
+            self._save_representations,
+        )
 
     def __call__(self, *args, **kwargs) -> 'Activity2':
         """Short-hand version of the `child` method."""
@@ -66,7 +73,7 @@ class Activity2(TreeObject, TagsMixin):
     def refresh(self, *args, **kwargs):
         """Refresh the object in place."""
         super().refresh(url=self._client._build_url('activity', activity_id=self.id),
-                        extra_params=API_EXTRA_PARAMS['activity'])
+                        extra_params=API_EXTRA_PARAMS['activity'], *args, **kwargs)
 
     #
     # additional properties
@@ -106,6 +113,19 @@ class Activity2(TreeObject, TagsMixin):
     @scope_id.setter
     def scope_id(self, value):
         self._scope_id = value
+
+    @property
+    def representations(self):
+        """Get and set the activity representations."""
+        return self._representations_container.get_representations()
+
+    @representations.setter
+    def representations(self, value):
+        self._representations_container.set_representations(value)
+
+    def _save_representations(self, representation_options):
+        self._options.update({'representations': representation_options})
+        self.edit(activity_options=self._options)
 
     #
     # predicates
@@ -444,6 +464,7 @@ class Activity2(TreeObject, TagsMixin):
             assignees_ids: Optional[List[Text]] = None,
             status: Optional[Union[ActivityStatus, Text]] = None,
             tags: Optional[List[Text]] = None,
+            **kwargs
     ) -> None:
         """Edit the details of an activity.
 
@@ -515,18 +536,19 @@ class Activity2(TreeObject, TagsMixin):
             assignees=assignees,
             assignees_ids=assignees_ids,
             status=status,
+            **kwargs,
         )
 
         url = self._client._build_url('activity', activity_id=self.id)
 
-        response = self._client._request('PUT', url, json=update_dict)
+        response = self._client._request('PUT', url, json=update_dict, params=API_EXTRA_PARAMS['activity'])
 
         if response.status_code != requests.codes.ok:  # pragma: no cover
             raise APIError("Could not update Activity {}".format(self), response=response)
 
         self.refresh(json=response.json().get('results')[0])
 
-    def _validate_edit_arguments(self, update_dict, start_date, due_date, assignees, assignees_ids, status):
+    def _validate_edit_arguments(self, update_dict, start_date, due_date, assignees, assignees_ids, status, **kwargs):
         """Verify inputs provided in both the `edit` and `edit_cascade_down` methods."""
         if assignees and assignees_ids:
             raise IllegalArgumentError('Provide either assignee names or their ids, but not both.')
@@ -551,6 +573,9 @@ class Activity2(TreeObject, TagsMixin):
             'due_date': check_datetime(dt=due_date, key='due_date'),
             'status': check_enum(status, ActivityStatus, 'status') or self.status,
         })
+
+        if kwargs:
+            update_dict.update(kwargs)
 
         return update_dict